+++ src/OrdersList/OrdersList.test.js
@@ -33,6 +33,7 @@ jest.mock('@folio/stripes-acq-components', () => ({
   useLocalStorageFilters: jest.fn(() => Object.values(mockLocalStorageFilters)),
   useFiltersToogle: jest.fn().mockReturnValue({ isFiltersOpened: true, toggleFilters: jest.fn() }),
   useModalToggle: jest.fn().mockReturnValue([false, jest.fn((state) => !state)]),
+  useItemToView: jest.fn().mockReturnValue({}),
   ResetButton: () => <span>ResetButton</span>,
   SingleSearchForm: () => <span>SingleSearchForm</span>,
 }));