+++ src/OrdersList/OrdersListContainer.js
@@ -1,14 +1,10 @@
-import React, { useCallback, useState } from 'react';
+import React, { useCallback } from 'react';
 import PropTypes from 'prop-types';
-import { withRouter } from 'react-router-dom';
-import ReactRouterPropTypes from 'react-router-prop-types';
-import queryString from 'query-string';
 
-import { getFullName } from '@folio/stripes/util';
 import { stripesConnect } from '@folio/stripes/core';
 import {
   organizationsManifest,
-  useList,
+  usePagination,
 } from '@folio/stripes-acq-components';
 
 import { RESULT_COUNT_INCREMENT } from '../common/constants';
@@ -23,103 +19,57 @@ import {
   fetchOrderUsers,
   fetchOrderVendors,
 } from './utils';
-import useBuildQuery from './useBuildQuery';
+
+import {
+  useOrders,
+} from './hooks';
 
 const resetData = () => { };
 
-const OrdersListContainer = ({ mutator, location }) => {
-  const [vendorsMap, setVendorsMap] = useState({});
-  const [acqUnitsMap, setAcqUnitsMap] = useState({});
-  const [usersMap, setUsersMap] = useState({});
-  const [ordersQuery, setOrdersQuery] = useState();
-  const buildQuery = useBuildQuery();
-
-  const loadOrders = useCallback(async (offset) => {
-    const query = buildQuery(queryString.parse(location.search));
-
-    setOrdersQuery(query);
-
-    return mutator.ordersListRecords.GET({
-      params: {
-        limit: RESULT_COUNT_INCREMENT,
-        offset,
-        query,
-      },
-    });
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [location.search]);
-
-  const loadOrdersCB = useCallback((setOrders, ordersResponse) => {
-    const fetchVendorsPromise = fetchOrderVendors(
-      mutator.orderVendors, ordersResponse.purchaseOrders, vendorsMap,
-    );
-    const fetchAcqUnitsPromise = fetchOrderAcqUnits(
-      mutator.orderAcqUnits, ordersResponse.purchaseOrders, acqUnitsMap,
-    );
-    const fetchUsersPromise = fetchOrderUsers(mutator.orderUsers, ordersResponse.purchaseOrders, usersMap);
+const OrdersListContainer = ({ mutator }) => {
+  const fetchReferences = useCallback(purchaseOrders => {
+    const fetchVendorsPromise = fetchOrderVendors(mutator.orderVendors, purchaseOrders, {});
+    const fetchAcqUnitsPromise = fetchOrderAcqUnits(mutator.orderAcqUnits, purchaseOrders, {});
+    const fetchUsersPromise = fetchOrderUsers(mutator.orderUsers, purchaseOrders, {});
 
     return Promise.all([fetchVendorsPromise, fetchAcqUnitsPromise, fetchUsersPromise])
       .then(([vendorsResponse, acqUnitsResponse, usersResponse]) => {
-        const newVendorsMap = {
-          ...vendorsMap,
-          ...vendorsResponse.reduce((acc, vendor) => {
-            acc[vendor.id] = vendor;
+        const vendorsMap = vendorsResponse.reduce((acc, vendor) => {
+          acc[vendor.id] = vendor;
 
-            return acc;
-          }, {}),
-        };
+          return acc;
+        }, {});
 
-        const newAcqUnitsMap = {
-          ...acqUnitsMap,
-          ...acqUnitsResponse.reduce((acc, unit) => {
-            acc[unit.id] = unit;
+        const acqUnitsMap = acqUnitsResponse.reduce((acc, unit) => {
+          acc[unit.id] = unit;
 
-            return acc;
-          }, {}),
-        };
+          return acc;
+        }, {});
 
-        const newUsersMap = {
-          ...usersMap,
-          ...usersResponse.reduce((acc, user) => {
-            acc[user.id] = user;
+        const usersMap = usersResponse.reduce((acc, user) => {
+          acc[user.id] = user;
 
-            return acc;
-          }, {}),
-        };
+          return acc;
+        }, {});
 
-        setVendorsMap(newVendorsMap);
-        setAcqUnitsMap(newAcqUnitsMap);
-        setUsersMap(newUsersMap);
-        setOrders((prev) => [
-          ...prev,
-          ...ordersResponse.purchaseOrders.map(order => ({
-            ...order,
-            vendorCode: newVendorsMap[order.vendor]?.code,
-            acquisitionsUnit: order.acqUnitIds?.map(unitId => newAcqUnitsMap[unitId]?.name).filter(Boolean).join(', '),
-            assignedTo: getFullName(newUsersMap[order.assignedTo]),
-          })),
-        ]);
+        return { usersMap, acqUnitsMap, vendorsMap };
       });
   // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [acqUnitsMap, usersMap, vendorsMap]);
+  }, []);
 
-  const {
-    records: orders,
-    recordsCount: ordersCount,
-    isLoading,
-    onNeedMoreData,
-    refreshList,
-  } = useList(false, loadOrders, loadOrdersCB, RESULT_COUNT_INCREMENT);
+  const { pagination, changePage, refreshPage } = usePagination({ limit: RESULT_COUNT_INCREMENT, offset: 0 });
+  const { query, orders, isLoading, ordersCount } = useOrders({ pagination, fetchReferences });
 
   return (
     <OrdersList
       ordersCount={ordersCount}
       isLoading={isLoading}
-      onNeedMoreData={onNeedMoreData}
+      onNeedMoreData={changePage}
       orders={orders}
-      refreshList={refreshList}
+      pagination={pagination}
+      refreshList={refreshPage}
       resetData={resetData}
-      ordersQuery={ordersQuery}
+      ordersQuery={query}
     />
   );
 };
@@ -147,8 +97,7 @@ OrdersListContainer.manifest = Object.freeze({
 });
 
 OrdersListContainer.propTypes = {
-  location: ReactRouterPropTypes.location.isRequired,
   mutator: PropTypes.object.isRequired,
 };
 
-export default withRouter(stripesConnect(OrdersListContainer));
+export default stripesConnect(OrdersListContainer);