+++ src/OrdersList/OrdersListContainer.test.js
@@ -2,21 +2,15 @@ import React from 'react';
 import { act, render, screen } from '@testing-library/react';
 import { MemoryRouter } from 'react-router';
 
-import {
-  useList,
-} from '@folio/stripes-acq-components';
-
 import { order } from '../../test/jest/fixtures';
 import { location } from '../../test/jest/routerMocks';
 
 import OrdersListContainer from './OrdersListContainer';
 import OrdersList from './OrdersList';
+import { useOrders } from './hooks';
 
 const defaultProps = {
   mutator: {
-    ordersListRecords: {
-      GET: jest.fn(),
-    },
     orderVendors: {
       GET: jest.fn().mockResolvedValue([{ id: 'vendorId' }]),
     },
@@ -38,29 +32,16 @@ const renderOrdersListContainer = (props = {}) => render(
   { wrapper: MemoryRouter },
 );
 
-const mockLocalStorageFilters = {
-  filters: {},
-  searchQuery: '',
-  applyFilters: jest.fn(),
-  applySearch: jest.fn(),
-  changeSearch: jest.fn(),
-  resetFilters: jest.fn(),
-  changeIndex: jest.fn(),
-  searchIndex: '',
-};
-
-jest.mock('@folio/stripes-acq-components', () => ({
-  ...jest.requireActual('@folio/stripes-acq-components'),
-  useLocalStorageFilters: jest.fn(() => Object.values(mockLocalStorageFilters)),
-  useFiltersToogle: jest.fn().mockReturnValue({ isFiltersOpened: true, toggleFilters: jest.fn() }),
-  useList: jest.fn().mockReturnValue({}),
-}));
 jest.mock('./OrdersList', () => jest.fn().mockReturnValue('OrdersList'));
+jest.mock('./hooks', () => ({
+  ...jest.requireActual('./hooks'),
+  useOrders: jest.fn().mockReturnValue({}),
+}));
 
 describe('OrdersListContainer', () => {
   beforeEach(() => {
-    useList.mockClear();
-    defaultProps.mutator.ordersListRecords.GET.mockClear();
+    useOrders.mockClear();
+
     defaultProps.mutator.orderVendors.GET.mockClear();
     defaultProps.mutator.orderAcqUnits.GET.mockClear();
     defaultProps.mutator.orderUsers.GET.mockClear();
@@ -72,39 +53,20 @@ describe('OrdersListContainer', () => {
     expect(screen.getByText('OrdersList')).toBeInTheDocument();
   });
 
-  it('should pass useList result to OrdersList', () => {
+  it('should pass useOrders result to OrdersList', () => {
     const records = [order];
 
     OrdersList.mockClear();
-    useList.mockReturnValue({ records });
+    useOrders.mockReturnValue({ orders: records });
     renderOrdersListContainer();
 
     expect(OrdersList.mock.calls[0][0].orders).toBe(records);
   });
 
-  it('should load orders in useList', async () => {
-    const OFFSET = 0;
-
+  it('should load vendors, order acq ids and users when fetchReferences is called', async () => {
     renderOrdersListContainer();
 
-    await act(() => useList.mock.calls[0][1](OFFSET));
-
-    expect(defaultProps.mutator.ordersListRecords.GET).toHaveBeenCalledWith({
-      params: {
-        limit: 30,
-        offset: OFFSET,
-        query: '(cql.allRecords=1) sortby metadata.updatedDate/sort.descending',
-      },
-    });
-  });
-
-  it('should load vendors, order acq ids and users if exist in useList', async () => {
-    const ordersResponse = { purchaseOrders: [order], totalRecords: 1 };
-    const setOrders = jest.fn();
-
-    renderOrdersListContainer();
-
-    await act(() => useList.mock.calls[0][2](setOrders, ordersResponse));
+    await act(() => useOrders.mock.calls[0][0].fetchReferences([order]));
 
     expect(defaultProps.mutator.orderVendors.GET).toHaveBeenCalledWith({
       params: {
@@ -126,7 +88,5 @@ describe('OrdersListContainer', () => {
         query: `id==${order.assignedTo}`,
       },
     });
-
-    expect(setOrders).toHaveBeenCalled();
   });
 });