+++ src/OrdersList/hooks/useBuildQuery/useBuildQuery.test.js
@@ -1,7 +1,7 @@
 import queryString from 'query-string';
 import { renderHook } from '@testing-library/react-hooks';
 
-import useBuildQuery from './useBuildQuery';
+import { useBuildQuery } from './useBuildQuery';
 
 describe('useBuildQuery', () => {
   it('should return function, that return query', () => {