+++ test/bigtest/network/factories/note-type.js
@@ -2,13 +2,13 @@ import { Factory } from 'miragejs';
 import faker from 'faker';
 
 export default Factory.extend({
-  id: faker.random.uuid,
+  id: faker.datatype.uuid,
   name: faker.commerce.productName,
   metadata: {
-    createdByUserId: faker.random.uuid,
+    createdByUserId: faker.datatype.uuid,
     createdByUsername: faker.name.firstName,
     createdDate: () => faker.date.past(2),
-    updatedByUserId: faker.random.uuid,
+    updatedByUserId: faker.datatype.uuid,
     updatedDate: () => faker.date.past(1),
   },
   usage: {