+++ test/bigtest/network/factories/orderTemplate.js
@@ -2,7 +2,7 @@ import { Factory } from 'miragejs';
 import faker from 'faker';
 
 export default Factory.extend({
-  id: faker.random.uuid,
+  id: faker.datatype.uuid,
   templateName: faker.finance.accountName,
   templateCode: faker.finance.accountName,
   orderFormat: () => 'Other',