+++ test/bigtest/network/factories/receivingHistory.js
@@ -2,8 +2,8 @@ import { Factory } from 'miragejs';
 import faker from 'faker';
 
 export default Factory.extend({
-  id: () => faker.random.uuid(),
-  poLineId: () => faker.random.uuid(),
+  id: () => faker.datatype.uuid(),
+  poLineId: () => faker.datatype.uuid(),
   poLineNumber: (id) => `${id}${faker.random.alphaNumeric()}${faker.random.alphaNumeric()}${faker.random.alphaNumeric()}`,
   title: (id) => `Piece - ${id}`,
   dateOrdered: () => faker.date.past(),